fix(rooms): change field dailyRate type from Double to BigDecimal

+++ src/main/java/tcc/caioferraz/backendapi/domain/rooms/RoomModel.java
@@ -3,6 +3,7 @@ package tcc.caioferraz.backendapi.domain.rooms;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -32,7 +33,7 @@ public class RoomModel {
   private String description;
 
   @Column(unique = true, nullable = false)
-  private Double dailyRate;
+  private BigDecimal dailyRate;
 
   @Column(unique = true, nullable = false)
   private Boolean isOcuped;
